xen: set dedicated vcpu to a cpu
When a VCPU is set to be dedicated VCPU, the VCPU will own that cpu and
the scheduler on that dedicated cpu will be disabled.
Benefit: This will avoid the unnecessary schedule overhead on the
dedicated cpu, which could benefit the extreme low latency applications.

NB: This version added trace mechanism for SCHEDULE_SOFTIRQ, scheduler choose a
fast path for the dedicated cpu.
NB: vcpu_block() is called super frequently on the dedicated cpu once we by-pass the
scheduler on the dedicated cpu. Not sure why this happens. Need to ask
in Xen ML.

Signed-off-by: Meng Xu <mengxu@cis.upenn.edu>