xen:rtds:Fix bug in budget accounting
Bug scenario:
While a VCPU is running on a core, it may misses its deadline.
Then repl_timer_handler() will update the VCPU period and deadline.
The VCPU may have high priority with the new deadline and repl_timer_handler()
decides to keep the VCPU running on the core, but with new period and deadline.
However, the budget enforcement timer for the previous period is still armed.
When rt_schedule() is called in the new period to enforce the budget
for the previous period, current burn_budget() will deduct the time spent
in previous period from the budget in current period, which is incorrect.

Fix:
We keeps last_start always within the current period for a VCPU, so that
We only deduct the time spent in the current period from the VCPU budget.

Signed-off-by: Meng Xu <mengxu@cis.upenn.edu>
Reported-by: Dagaen Golomb <dgolomb@cis.upenn.edu>
---
Cc: Dario Faggioli <dario.faggioli@citrix.com>
Cc: George Dunlap <George.Dunlap@eu.citrix.com>
Cc: Wei Liu <wei.liu2@citrix.com>
Cc: Linh Thi Xuan Phan <linhphan@cis.upenn.edu>
Cc: Haoran Li <lihaoran@wustl.edu>
Cc: Meng Xu <xumengpanda@gmail.com>
Cc: Dagaen Golomb <dgolomb@cis.upenn.edu>
Cc: Tianyang Chen <tiche@cis.upenn.edu>
---
Changes from v1:
* Change commit message to make the bug scenario easire to understand;
* The two bug scenarios described in v1 can be actually fixed by this patch;
  so we do not need to change the runq_tickle